Solaris gcc -m64 fixes
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@11930 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/agent/extend.c
@@ -24,7 +24,7 @@ typedef struct extend_registration_block_s {
     netsnmp_table_data *dinfo;
     oid                *root_oid;
     size_t              oid_len;
-    int                 num_entries;
+    long                num_entries;
     netsnmp_extend     *ehead;
     struct extend_registration_block_s *next;
 } extend_registration_block;
